+++ src/main/java/cc/cassian/raspberry/mixin/RaspberryMixinCanceller.java
@@ -1,6 +1,7 @@
 package cc.cassian.raspberry.mixin;
 
 import com.bawnorton.mixinsquared.api.MixinCanceller;
+import net.minecraftforge.fml.loading.LoadingModList;
 
 import java.util.List;
 
@@ -10,7 +11,7 @@ public class RaspberryMixinCanceller implements MixinCanceller {
         if (mixinClassName.equals("me.juancarloscp52.spyglass_improvements.mixin.MinecraftMixin")) {
             return true;
         }
-        if (mixinClassName.equals("com.mrcrayfish.controllable.mixin.client.jei.IngredientListOverlayMixin")) {
+        else if (mixinClassName.equals("com.mrcrayfish.controllable.mixin.client.jei.IngredientListOverlayMixin") && LoadingModList.get().getModFileById("emi") != null) {
             return true;
         }
         return false;