+++ src/main/java/cc/cassian/raspberry/config/ModConfig.java
@@ -42,6 +42,7 @@ public class ModConfig {
     public int mirrorParticleSearchRadius = 48;
     public int mirrorVerticalParticleSearchRadius = 24;
     public double mirrorVolumeModifier = 0.4;
+    public boolean sunSensitiveRaiders = false;
 
     public static void load() {
         if (!Files.exists(configPath())) {