+++ src/main/java/cc/cassian/raspberry/mixin/minecraft/RaiderMixin.java
@@ -1,0 +1,47 @@
+package cc.cassian.raspberry.mixin.minecraft;
+
+import cc.cassian.raspberry.config.ModConfig;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.monster.PatrollingMonster;
+import net.minecraft.world.entity.raid.Raider;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(Raider.class)
+public abstract class RaiderMixin extends PatrollingMonster {
+    protected RaiderMixin(EntityType<? extends PatrollingMonster> entityType, Level level) {
+        super(entityType, level);
+    }
+
+    @Inject(method = "aiStep", at = @At(value = "HEAD"))
+    public void burnInDaylight(CallbackInfo ci) {
+        if (this.isAlive()) {
+            boolean flag = ModConfig.get().sunSensitiveRaiders && this.isSunBurnTick();
+            if (flag) {
+                ItemStack itemstack = this.getItemBySlot(EquipmentSlot.HEAD);
+                if (!itemstack.isEmpty()) {
+                    if (itemstack.isDamageableItem()) {
+                        itemstack.setDamageValue(itemstack.getDamageValue() + this.random.nextInt(2));
+                        if (itemstack.getDamageValue() >= itemstack.getMaxDamage()) {
+                            this.broadcastBreakEvent(EquipmentSlot.HEAD);
+                            this.setItemSlot(EquipmentSlot.HEAD, ItemStack.EMPTY);
+                        }
+                    }
+
+                    flag = false;
+                }
+
+                if (flag) {
+                    this.setSecondsOnFire(8);
+                }
+            }
+        }
+
+        super.aiStep();
+    }
+}