+++ src/main/java/cc/cassian/raspberry/config/ModConfig.java
@@ -43,6 +43,7 @@ public class ModConfig {
     public int mirrorVerticalParticleSearchRadius = 24;
     public double mirrorVolumeModifier = 0.4;
     public boolean sunSensitiveRaiders = false;
+    public boolean mapsWorkInInventory = false;
 
     public static void load() {
         if (!Files.exists(configPath())) {