+++ src/main/java/cc/cassian/raspberry/mixin/minecraft/MapItemMixin.java
@@ -1,0 +1,21 @@
+package cc.cassian.raspberry.mixin.minecraft;
+
+import cc.cassian.raspberry.config.ModConfig;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import com.llamalad7.mixinextras.sugar.Local;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.MapItem;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin(MapItem.class)
+public class MapItemMixin {
+    @WrapOperation(method = "inventoryTick", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/player/Player;getOffhandItem()Lnet/minecraft/world/item/ItemStack;"))
+    private ItemStack mixin(Player instance, Operation<ItemStack> original, @Local(argsOnly = true) ItemStack stack) {
+        if (ModConfig.get().mapsWorkInInventory) {
+            return stack;
+        } else return original.call(instance);
+    }
+}