+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Cartographer's Quills not having names set correctly. (Supplementaries)
 - Respiration tablet recipe using Iron Helmets instead of the helmet tag (EMI, Raspberry Flavoured)
+- Disabled Controllable JEI compatibility mixin when EMI is present.
 
 ## [1.6] - 2025-05-19