+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cartographer's Quills not having names set correctly. (Supplementaries)
 - Respiration tablet recipe using Iron Helmets instead of the helmet tag (EMI, Raspberry Flavoured)
 - Disabled Controllable JEI compatibility mixin when EMI is present.
+- Ashballs can now be dispensed.
 
 ## [1.6] - 2025-05-19