+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Backported Auditory Continued sounds for various eating sounds.
 - Backport of Sage Brush to 1.19.2.
 - Raiders now burn in daylight. (disabled by default)
+- Maps now work in the inventory (disabled by default)
 
 ### Changed
 - Knight Jump can no longer be used in water, preventing accidental durability loss. (Mining Master)